extract adding streams to new file
